Make collision rect-only
This huge commit implements diagonal collision sliding, but it forces the physics to use only box-shaped collisions.

The only slight bug left of collision (in my opinion) is the fact that fast speeds create bigger gaps when the collision stops the movement. For instance: when you dash-slash towards a wall, you will stop much farther from the wall had you just walked up to it.
Fixing this will require lots of refactoring, I assume. I am considering changing the game units so that the tilemap's units are in pixels. This makes collision less float-based, and could potentially help 'finding' those positions that are closest before collision, and thus fixing the high-speed-high-gap problem.

But this is low priority now.
Closes #13